requirements.txt
# Core dependencies
claude-agent-sdk>=0.1.0
anthropic>=0.40.0

# YouTube API and transcript extraction
youtube-transcript-api>=0.6.0
google-api-python-client>=2.0.0

# Utilities
python-dotenv>=1.0.0
anyio>=4.0.0
